random_stuff/requirements.txt
discord
python-dotenv
pytz
firebase_admin
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib